requirements.txt
fastapi
uvicorn
transformers
torch
sentencepiece
pandas
plotly
scikit-learn
faiss-cpu
langchain
openai